requirements.txt
flask
mysql-connector-python==9.3.0
tensorflow
numpy
flask_cors
wordcloud
nltk